refactor: reorganize imports and move _ExporterManager class and functions to pyquerytracker/exporter/utils.py for better structure
